requirements.txt
attrs==17.4.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
clickclick==1.2.2
connexion==1.3
flask==0.12.2
flask-script==2.0.6
gunicorn==19.7.1
idna==2.6
inflection==0.3.1
itsdangerous==0.24
jinja2==2.10
jsonschema==2.6.0
markupsafe==1.0
pyyaml==3.12
requests==2.18.4
six==1.11.0
swagger-spec-validator==2.1.0
urllib3==1.22
werkzeug==0.14.1
